Moved tests in IntegrationTest according to production structure
