+++ tests/LinkDotNet.Blog.IntegrationTests/Web/Features/Admin/Dashboard/Components/VisitCountPerPageTests.cs
@@ -9,7 +9,7 @@ using LinkDotNet.Blog.Web.Features.Admin.Dashboard.Components;
 using LinkDotNet.Blog.Web.Features.Admin.Dashboard.Services;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace LinkDotNet.Blog.IntegrationTests.Web.Pages.Admin.Dashboard;
+namespace LinkDotNet.Blog.IntegrationTests.Web.Features.Admin.Dashboard.Components;
 
 public class VisitCountPerPageTests : SqlDatabaseTestBase<BlogPost>
 {
